requirements.txt
Flask==2.3.3
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
timm==0.9.7
Werkzeug==2.3.7
gunicorn==21.2.0
psutil==5.9.5
